fix: Correct (generated) types structure
