+++ tsconfig.json
@@ -14,7 +14,7 @@
      "declaration": true,                   /* Generates corresponding '.d.ts' file. */
     // "declarationMap": true,                /* Generates a sourcemap for each corresponding '.d.ts' file. */
     // "sourceMap": true,                     /* Generates corresponding '.map' file. */
-     "outFile": "./index.d.ts",                       /* Concatenate and emit output to single file. */
+    // "outFile": "./index.d.ts",                       /* Concatenate and emit output to single file. */
     // "outDir": ".",                        /* Redirect output structure to the directory. */
     // "rootDir": "./src/",                       /* Specify the root directory of input files. Use to control the output directory structure with --outDir. */
     // "composite": true,                     /* Enable project compilation */
